Only need the right hand side of the ensures clause for function calls
